Avoid creating an array with 0 elements
In buffers returned by XPAGet, bufs[i] can be non-NULL while lens[i] can
be equal to 0.  In that case, we return an empty result (or a nil
string) rather than an array of bytes with 0 elements.  In <yapi.h>, you
can create such an array (but you will get into trouble if you print its
contents) but this is forbidden in Yorick interpreter.